requirements.txt
matplotlib
numpy<2
Pillow
torchinfo
pandas
networkx
scipy
torch
torchvision
torch-scatter
torch-geometric
scikit-image
tqdm